refactor(import-compliance): complete P17 - remove all re-exports and relative imports
- Fix relative imports in configuration/cli.py to use absolute imports
- Update all consumers to import directly from canonical submodules instead of through __init__.py re-exports
- Remove re-exports from training, sampling, configuration, and data_pipeline package __init__.py files
- Update CLI, experiments, and tests to use direct submodule imports
- Achieve 100% compliance with IMPORT_GUIDELINES.md by eliminating all re-export facades
- Ensure zero backwards compatibility - all imports must be from canonical locations

All quality gates pass: linting, type checking, and full test suite (221 passed, 5 skipped).
