+++ ml_playground/data_pipeline/sampling/__init__.py
@@ -1,5 +1,1 @@
 """Sampling utilities for data pipelines."""
-
-from ml_playground.data_pipeline.sampling.batches import SimpleBatches, sample_batch
-
-__all__ = ["SimpleBatches", "sample_batch"]
