+++ ml_playground/ema.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 
-from ml_playground.configuration import DeviceKind
+from ml_playground.configuration.models import DeviceKind
 from ml_playground.models.core.model import GPT
