+++ ml_playground/experiments/bundestag_qwen15b_lora_mps/preparer.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Iterable
 
-from ml_playground.configuration import PreparerConfig
+from ml_playground.configuration.models import PreparerConfig
 from ml_playground.experiments.protocol import (
     Preparer as _PreparerProto,
     PrepareReport,
