+++ ml_playground/experiments/bundestag_qwen15b_lora_mps/sampler.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from pathlib import Path
-from ml_playground.configuration import SamplerConfig
+from ml_playground.configuration.models import SamplerConfig
 from ml_playground.experiments.protocol import (
     Sampler as _SamplerProto,
     SampleReport,
