+++ ml_playground/experiments/shakespeare/preparer.py
@@ -4,12 +4,14 @@ from pathlib import Path
 import numpy as np
 import requests
 import requests.exceptions
-from ml_playground.configuration import PreparerConfig
-from ml_playground.data_pipeline import (
+from ml_playground.configuration.models import PreparerConfig
+from ml_playground.data_pipeline.transforms.tokenization import (
     create_standardized_metadata,
+    split_train_val,
+)
+from ml_playground.data_pipeline.transforms.io import (
     diff_file_states,
     snapshot_file_states,
-    split_train_val,
     write_bin_and_meta,
 )
 from ml_playground.tokenizer import create_tokenizer
