+++ ml_playground/experiments/speakger/trainer.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from pathlib import Path
-from ml_playground.configuration import TrainerConfig
+from ml_playground.configuration.models import TrainerConfig
 from ml_playground.experiments.protocol import (
     Trainer as _TrainerProto,
     TrainReport,
