+++ ml_playground/sampling/__init__.py
@@ -1,10 +1,3 @@
 """Sampling utilities for ml_playground."""
 
 from __future__ import annotations
-
-from ml_playground.sampling.runner import Sampler, sample
-
-__all__ = [
-    "Sampler",
-    "sample",
-]
