+++ ml_playground/sampling/runner.py
@@ -13,7 +13,7 @@ import torch
 from torch import autocast
 
 from ml_playground.checkpoint import Checkpoint, CheckpointManager
-from ml_playground.configuration import (
+from ml_playground.configuration.models import (
     ModelConfig,
     SamplerConfig,
     READ_POLICY_BEST,
