+++ ml_playground/training/__init__.py
@@ -1,5 +1,1 @@
 """Training orchestration package."""
-
-from ml_playground.training.loop.runner import Trainer, train, get_lr
-
-__all__ = ["Trainer", "train", "get_lr"]
