+++ ml_playground/training/checkpointing/__init__.py
@@ -1,17 +1,1 @@
 """Checkpointing helpers for the training package."""
-
-from .service import (
-    apply_checkpoint,
-    create_manager,
-    load_checkpoint,
-    propagate_metadata,
-    save_checkpoint,
-)
-
-__all__ = [
-    "apply_checkpoint",
-    "create_manager",
-    "load_checkpoint",
-    "propagate_metadata",
-    "save_checkpoint",
-]
