+++ ml_playground/training/hooks/components.py
@@ -8,7 +8,7 @@ import torch
 from torch.amp.grad_scaler import GradScaler
 from torch.utils.tensorboard import SummaryWriter
 
-from ml_playground.configuration import TrainerConfig
+from ml_playground.configuration.models import TrainerConfig
 from ml_playground.ema import EMA
 from ml_playground.models.core.model import GPT
 from ml_playground.training.hooks.runtime import RuntimeContext
