+++ ml_playground/training/hooks/data.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from ml_playground.configuration import TrainerConfig, SharedConfig
+from ml_playground.configuration.models import TrainerConfig, SharedConfig
 from ml_playground.data_pipeline.sampling.batches import SimpleBatches
