+++ ml_playground/training/loop/__init__.py
@@ -1,5 +1,1 @@
 """Loop-level orchestration utilities."""
-
-from .runner import Trainer, get_lr, train
-
-__all__ = ["Trainer", "get_lr", "train"]
