+++ tests/e2e/ml_playground/experiments/speakger/test_sampler_analysis.py
@@ -10,7 +10,11 @@ import pytest
 
 # Import the module under test
 from ml_playground.experiments.speakger import sampler as gm
-from ml_playground.configuration import SamplerConfig, RuntimeConfig, SampleConfig
+from ml_playground.configuration.models import (
+    SamplerConfig,
+    RuntimeConfig,
+    SampleConfig,
+)
 
 
 class DummyTokenizer:
