+++ tests/unit/training/checkpointing/test_service.py
@@ -6,7 +6,7 @@ from typing import Any
 import pytest
 import torch
 
-from ml_playground.configuration import (
+from ml_playground.configuration.models import (
     DataConfig,
     LRSchedule,
     ModelConfig,
