+++ tests/unit/training/loop/test_training_runner.py
@@ -8,7 +8,7 @@ import pytest
 import torch
 
 import ml_playground.training.loop.runner as runner_mod
-from ml_playground.configuration import (
+from ml_playground.configuration.models import (
     DataConfig,
     LRSchedule,
     ModelConfig,
